Replace setjmp-longjmp's by try_catch-throw   modified:   storage/connect/CMakeLists.txt   modified:   storage/connect/array.cpp   modified:   storage/connect/blkfil.cpp   modified:   storage/connect/colblk.cpp   modified:   storage/connect/connect.cc   modified:   storage/connect/filamtxt.cpp   modified:   storage/connect/filamvct.cpp   modified:   storage/connect/filter.cpp   modified:   storage/connect/global.h   modified:   storage/connect/ha_connect.cc   modified:   storage/connect/jdbconn.cpp   modified:   storage/connect/json.cpp   modified:   storage/connect/jsonudf.cpp   modified:   storage/connect/odbconn.cpp   modified:   storage/connect/osutil.c   modified:   storage/connect/plgdbutl.cpp   deleted:    storage/connect/plugutil.c   added:      storage/connect/plugutil.cpp   modified:   storage/connect/tabdos.cpp   modified:   storage/connect/tabfix.cpp   modified:   storage/connect/tabfmt.cpp   modified:   storage/connect/tabjdbc.cpp   modified:   storage/connect/tabjdbc.h   modified:   storage/connect/tabjson.cpp   modified:   storage/connect/tabmul.cpp   modified:   storage/connect/tabmul.h   modified:   storage/connect/tabmysql.cpp   modified:   storage/connect/tabodbc.cpp   modified:   storage/connect/tabodbc.h   modified:   storage/connect/tabpivot.cpp   modified:   storage/connect/tabsys.cpp   modified:   storage/connect/tabvct.cpp   modified:   storage/connect/tabvir.cpp   modified:   storage/connect/tabxml.cpp   modified:   storage/connect/valblk.cpp   modified:   storage/connect/value.cpp   modified:   storage/connect/xindex.cpp   modified:   storage/connect/xobject.cpp

+++ storage/connect/tabmul.h
@@ -134,7 +134,7 @@ class TDBDIR : public TDBASE {
   int  iFile;                   // Index of currently retrieved file
 #if defined(__WIN__)
   _finddata_t    FileData;      // Find data structure
-  int  Hsearch;                 // Search handle
+  intptr_t Hsearch;             // Search handle
   char Drive[_MAX_DRIVE];       // Drive name
 #else   // !__WIN__
   struct stat    Fileinfo;      // File info structure
@@ -184,7 +184,7 @@ class TDBSDR : public TDBDIR {
     struct _Sub_Dir *Next;
     struct _Sub_Dir *Prev;
 #if defined(__WIN__)
-    int H;               // Search handle
+    intptr_t H;               // Search handle
 #else   // !__WIN__
     DIR *D;
 #endif  // !__WIN__
